feat: 组织架构侧边栏 # Reviewed, transaction id: 6405
