+++ src/pages/src/language/en.json
@@ -450,5 +450,7 @@
   "时间": "Time",
   "更新内容": "Update content",
   "变更": "Change",
+  "添加子组织": "Add sub-organization",
+  "重命名": "Rename",
   "新增": "Add"
 }
