+++ src/pages/src/language/zh.json
@@ -450,5 +450,7 @@
   "时间": "时间",
   "更新内容": "更新内容",
   "变更": "变更",
+  "添加子组织": "添加子组织",
+  "重命名": "重命名",
   "新增": "新增"
 }
