+++ src/pages/src/types/organization.d.ts
@@ -1,0 +1,7 @@
+export interface IOrg {
+  id: number;
+  name: string;
+  has_children: boolean;
+  async: boolean;
+  children?: IOrg[];
+}
